requirements.txt
amqp==5.1.1
asgiref==3.5.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.1
cached-property==1.5.2
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.14
django-appconf==1.0.5
django-audiofield==0.10.0
django-auto-logout==0.5.0
django-celery==3.1.17
django-cors-headers==3.7.0
django-debug-toolbar==2.0
django-environ==0.8.1
django-filter==21.1
django-imagekit==4.1.0
django-jazzmin==2.4.9
django-js-asset==2.0.0
django-modeltranslation==0.17.3
django-mptt==0.13.4
django-solo==2.0.0
django-summernote==0.8.20.0
django-tinymce==3.4.0
djangorestframework==3.12.4
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
pilkit==2.0
Pillow==8.2.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==6.1.1
pytz==2022.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
text-unidecode==1.3
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
XlsxWriter==3.0.3
zipp==3.8.0
